refactor: cleaned and added some comments
